Check if bios is locked

+++ er2-disk.sh
@@ -17,11 +17,15 @@ clear='\033[0m'
 if dmidecode -s system-manufacturer | grep -qi "dell"; then
   mode=$(cctk --EmbSataRaid)
   if [[ "$mode" == "EmbSataRaid=Raid" ]]; then
-      cctk --EmbSataRaid=Ahci >> /dev/null
-      echo "Reboot required for script to continue! Restarting in 3 seconds..."
-      sleep 3
-      systemctl reboot
-      exit
+      sata_option=$(cctk --EmbSataRaid=Ahci | grep -oh "Setup Password")
+      if [[ "$sata_option" == "Setup Password" ]]; then
+        echo -e "${yellow}WARNING: BIOS Locked. Unable to update bios settings. The script may not be able to detect NVME drives in the device. Please shutdown the device and ensure there are no NVME drives installed before continuing.${clear}"
+      else
+        echo "Reboot required for script to continue! Restarting in 3 seconds..."
+        sleep 3
+        systemctl reboot
+        exit
+      fi
   fi
 fi
